requirements.txt
Flask==2.3.2
openai
regex==2023.6.3
requests==2.31.0
tiktoken==0.7.0
Werkzeug==2.3.6
Markdown==3.4.4
flask_cors
python-dotenv
gunicorn
camel-ai==0.1.5.2
Pillow
unstructured
nanoid
cachetools
loguru
piexif
pycryptodome